# Backend/TheLionKing/src/main/java/de/drv/thelionking/workflow/dto/PageInfoDto.java
package de.drv.thelionking.workflow.dto;

import java.util.UUID;

public class PageInfoDto {
    private UUID pageId;
    private int pageNo;
    private String status;
    private String errorMessage;

    public UUID getPageId() {
        return pageId;
    }

    public void setPageId(UUID pageId) {
        this.pageId = pageId;
    }

    public int getPageNo() {
        return pageNo;
    }

    public void setPageNo(int pageNo) {
        this.pageNo = pageNo;
    }

    public String getStatus() {
        return status;
    }

    public void setStatus(String status) {
        this.status = status;
    }

    public String getErrorMessage() {
        return errorMessage;
    }

    public void setErrorMessage(String errorMessage) {
        this.errorMessage = errorMessage;
    }
}